pyproject.toml
[project]
name = "p9-rag-events"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.72.0",
    "faiss-cpu>=1.12.0",
    "jupyterlab>=4.4.10",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-mistralai>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "mistralai>=1.9.11",
    "numpy>=2.3.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "voyageai>=0.3.5",
]
